refactor: move data column helpers to utils
