+++ packages/beacon-node/src/db/repositories/dataColumnSidecarsArchive.ts
@@ -2,8 +2,8 @@ import {ChainForkConfig} from "@lodestar/config";
 import {Db, Repository} from "@lodestar/db";
 import {Slot} from "@lodestar/types";
 import {bytesToInt} from "@lodestar/utils";
+import {DataColumnSidecarsDbWrapper, dataColumnSidecarsDbWrapperSsz} from "../../util/dataColumns.js";
 import {Bucket, getBucketNameByValue} from "../buckets.js";
-import {DataColumnSidecarsDbWrapper, dataColumnSidecarsDbWrapperSsz} from "./dataColumnSidecars.js";
 
 /**
  * dataColumnSidecarsWrapper by slot
