+++ packages/beacon-node/src/util/blobs.ts
@@ -6,12 +6,10 @@ import {
   ForkName,
   KZG_COMMITMENTS_GINDEX,
   KZG_COMMITMENT_GINDEX0,
-  NUMBER_OF_COLUMNS,
   VERSIONED_HASH_VERSION_KZG,
 } from "@lodestar/params";
 import {signedBlockToSignedHeader} from "@lodestar/state-transition";
 import {BeaconBlockBody, SSZTypesFor, SignedBeaconBlock, deneb, fulu, ssz} from "@lodestar/types";
-import {kzg} from "./rustKzg.js";
 
 type VersionHash = Uint8Array;
 
@@ -60,53 +58,5 @@ export function computeBlobSidecars(
       contents.kzgCommitmentInclusionProofs?.[index] ?? computeInclusionProof(fork, signedBlock.message.body, index);
 
     return {index, blob, kzgCommitment, kzgProof, signedBlockHeader, kzgCommitmentInclusionProof};
-  });
-}
-
-/**
- * Turns a SignedBeaconBlock and an array of Blobs from a given slot into an array of
- * DataColumnSidecars that are ready to be served by gossip and req/resp.
- *
- * Implementation of get_data_column_sidecars
- * https://github.com/ethereum/consensus-specs/blob/dev/specs/_features/eip7594/das-core.md#get_data_column_sidecars
- */
-export function computeDataColumnSidecars(
-  config: ChainForkConfig,
-  signedBlock: SignedBeaconBlock,
-  contents: fulu.Contents & {kzgCommitmentsInclusionProof?: fulu.KzgCommitmentsInclusionProof; cells?: fulu.Cell[][]}
-): fulu.DataColumnSidecars {
-  const blobKzgCommitments = (signedBlock as deneb.SignedBeaconBlock).message.body.blobKzgCommitments;
-  if (blobKzgCommitments === undefined) {
-    throw Error("Invalid block with missing blobKzgCommitments for computeBlobSidecars");
-  }
-  if (blobKzgCommitments.length === 0) {
-    return [];
-  }
-  const fork = config.getForkName(signedBlock.message.slot);
-  const signedBlockHeader = signedBlockToSignedHeader(config, signedBlock);
-  const kzgCommitmentsInclusionProof =
-    contents.kzgCommitmentsInclusionProof ?? computeKzgCommitmentsInclusionProof(fork, signedBlock.message.body);
-  const {blobs, kzgProofs} = contents;
-  const cellsAndProofs = Array.from({length: blobs.length}, (_, rowNumber) => {
-    const cells = contents.cells?.[rowNumber] ?? kzg.computeCells(blobs[rowNumber]);
-    const proofs = kzgProofs.slice(rowNumber * NUMBER_OF_COLUMNS, (rowNumber + 1) * NUMBER_OF_COLUMNS);
-    return {cells, proofs};
-  });
-
-  return Array.from({length: NUMBER_OF_COLUMNS}, (_, columnIndex) => {
-    // columnIndex'th column
-    const column = Array.from({length: blobs.length}, (_, rowNumber) => cellsAndProofs[rowNumber].cells[columnIndex]);
-    const kzgProofs = Array.from(
-      {length: blobs.length},
-      (_, rowNumber) => cellsAndProofs[rowNumber].proofs[columnIndex]
-    );
-    return {
-      index: columnIndex,
-      column,
-      kzgCommitments: blobKzgCommitments,
-      kzgProofs,
-      signedBlockHeader,
-      kzgCommitmentsInclusionProof,
-    };
   });
 }
